Data.RAList.NonEmpty

Description

Non-empty random access list.

This module is designed to imported qualifed.

Documentation

newtype NERAList a

Non-empty random access list.

Constructors

NE (NERAList' Leaf a)

data NERAList' f a

Non-empty random access list, underlying representation.

The structure doesn't need to be hidden, as polymorphic recursion of Nodes starting from Leaf keeps the NERAList values well-formed.

Constructors

Last (f a)

Cons0 (NERAList' (Node f) a)

Cons1 (f a) (NERAList' (Node f) a)

Showing

explicitShow :: Show a => NERAList a -> String

explicitShowsPrec :: Show a => Int -> NERAList a -> ShowS

Construction

singleton :: a -> NERAList a

Single element NERAList.

cons :: a -> NERAList a -> NERAList a

cons for non-empty rals.

Indexing

(!) :: NERAList a -> Int -> a

List index.

>>> fromNonEmpty ('a' :| ['b'..'f']) ! 0
'a'

>>> fromNonEmpty ('a' :| ['b'..'f']) ! 5
'f'

>>> fromNonEmpty ('a' :| ['b'..'f']) ! 6
*** Exception: array index out of range: NERAList
...

(!?) :: NERAList a -> Int -> Maybe a

safe list index.

>>> fromNonEmpty ('a' :| ['b'..'f']) !? 0
Just 'a'

>>> fromNonEmpty ('a' :| ['b'..'f']) !? 5
Just 'f'

>>> fromNonEmpty ('a' :| ['b'..'f']) !? 6
Nothing

head :: NERAList a -> a

First value, head of the list.

>>> head $ fromNonEmpty $ 'a' :| ['b'..'f']
'a'

last :: NERAList a -> a

Last value of the list

>>> last $ fromNonEmpty $  'a' :| ['b'..'f']
'f'

uncons :: NERAList a -> (a, RAList a)

>>> uncons $ fromNonEmpty $ 'a' :| "bcdef"
('a',fromList "bcdef")

tail :: NERAList a -> RAList a

Tail of non-empty list can be empty.

>>> tail $ fromNonEmpty $ 'a' :| "bcdef"
fromList "bcdef"

length :: NERAList a -> Int

null :: NERAList a -> Bool

Conversions

toNonEmpty :: NERAList a -> NonEmpty a

fromNonEmpty :: NonEmpty a -> NERAList a

>>> fromNonEmpty ('a' :| ['b'..'f'])
fromNonEmpty ('a' :| "bcdef")

>>> explicitShow (fromNonEmpty ('a' :| ['b'..'f']))
"NE (Cons0 (Cons1 (Nd (Lf 'a') (Lf 'b')) (Last (Nd (Nd (Lf 'c') (Lf 'd')) (Nd (Lf 'e') (Lf 'f'))))))"

Folding

foldMap1 :: forall a s. Semigroup s => (a -> s) -> NERAList a -> s

foldr1Map :: (a -> b -> b) -> (a -> b) -> NERAList a -> b

ifoldMap :: Monoid m => (Int -> a -> m) -> NERAList a -> m

ifoldMap1 :: forall a s. Semigroup s => (Int -> a -> s) -> NERAList a -> s

>>> import Data.Semigroup (Min (..))

>>> ifoldMap1 (\_ x -> Min x) $ fromNonEmpty $ 5 :| [3,1,2,4]
Min {getMin = 1}

>>> ifoldMap1 (\i x -> Min (i + x)) $ fromNonEmpty $ 5 :| [3,1,2,4]
Min {getMin = 3}

ifoldr1Map :: forall a b. (Int -> a -> b -> b) -> (Int -> a -> b) -> NERAList a -> b

Mapping

adjust :: forall a. Int -> (a -> a) -> NERAList a -> NERAList a

Adjust a value in the list.

>>> adjust 3 toUpper $ fromNonEmpty $ 'a' :| "bcdef"
fromNonEmpty ('a' :| "bcDef")

If index is out of bounds, the list is returned unmodified.

>>> adjust 10 toUpper $ fromNonEmpty $ 'a' :| "bcdef"
fromNonEmpty ('a' :| "bcdef")

>>> adjust (-1) toUpper $ fromNonEmpty $ 'a' :| "bcdef"
fromNonEmpty ('a' :| "bcdef")

map :: (a -> b) -> NERAList a -> NERAList b

>>> map toUpper (fromNonEmpty ('a' :| ['b'..'f']))
fromNonEmpty ('A' :| "BCDEF")

imap :: (Int -> a -> b) -> NERAList a -> NERAList b

>>> imap (,) (fromNonEmpty ('a' :| ['b'..'f']))
fromNonEmpty ((0,'a') :| [(1,'b'),(2,'c'),(3,'d'),(4,'e'),(5,'f')])

itraverse :: forall f a b. Applicative f => (Int -> a -> f b) -> NERAList a -> f (NERAList b)

itraverse1 :: forall f a b. Apply f => (Int -> a -> f b) -> NERAList a -> f (NERAList b)