Data.RAList.Tree

Documentation

newtype Leaf a

A Leaf is isomorphic to Identity, but we reimplement it here to have domain specific type. The short constructor name is a bonus.

Constructors

Lf a

data Node f a

Node is a product of two f. This way we can form a perfect binary tree.

Constructors

Nd (f a) (f a)

data Dir a

Direction in Node.

Constructors

L a

R a