Data.RAVec.Tree

Description

Depth indexed perfect binary tree.

Documentation

data Tree (n :: Nat) a where

Perfectly balanced binary tree of depth n, with 2 ^ n elements.

Constructors

Leaf :: a -> Tree 'Z a

Node :: Tree n a -> Tree n a -> Tree ('S n) a

Construction

singleton :: a -> Tree 'Z a

Tree of zero depth, with single element.

>>> singleton True
Leaf True

Conversions

toList :: Tree n a -> [a]

Convert Tree to list.

>>> toList $ Node (Node (Leaf 'a') (Leaf 'b')) (Node (Leaf 'c') (Leaf 'd'))
"abcd"

reverse :: Tree n a -> Tree n a

Reverse Tree.

>>> let t = Node (Node (Leaf 'a') (Leaf 'b')) (Node (Leaf 'c') (Leaf 'd'))
>>> reverse t
Node (Node (Leaf 'd') (Leaf 'c')) (Node (Leaf 'b') (Leaf 'a'))

Since: 0.1.1

Indexing

(!) :: Tree n a -> Wrd n -> a

Indexing.

tabulate :: forall n a. SNatI n => (Wrd n -> a) -> Tree n a

leftmost :: Tree n a -> a

rightmost :: Tree n a -> a

Folds

foldMap :: Monoid m => (a -> m) -> Tree n a -> m

foldMap1 :: Semigroup s => (a -> s) -> Tree n a -> s

ifoldMap :: Monoid m => (Wrd n -> a -> m) -> Tree n a -> m

ifoldMap1 :: Semigroup s => (Wrd n -> a -> s) -> Tree n a -> s

foldr :: (a -> b -> b) -> b -> Tree n a -> b

>>> foldr (:) [] $ Node (Leaf True) (Leaf False)
[True,False]

ifoldr :: (Wrd n -> a -> b -> b) -> b -> Tree n a -> b

foldr1Map :: (a -> b -> b) -> (a -> b) -> Tree n a -> b

ifoldr1Map :: (Wrd n -> a -> b -> b) -> (Wrd n -> a -> b) -> Tree n a -> b

foldl :: (b -> a -> b) -> b -> Tree n a -> b

>>> foldl (flip (:)) [] $ Node (Leaf True) (Leaf False)
[False,True]

ifoldl :: (Wrd n -> b -> a -> b) -> b -> Tree n a -> b

Special folds

length :: Tree n a -> Int

>>> length (universe :: Tree N.Nat3 (Wrd N.Nat3))
8

null :: Tree n a -> Bool

Since: 0.1.1

sum :: Num a => Tree n a -> a

Non-strict sum.

Since: 0.1.1

product :: Num a => Tree n a -> a

Non-strict product.

Since: 0.1.1

Mapping

map :: (a -> b) -> Tree n a -> Tree n b

>>> map not $ Node (Leaf True) (Leaf False)
Node (Leaf False) (Leaf True)

imap :: (Wrd n -> a -> b) -> Tree n a -> Tree n b

>>> imap (,) $ Node (Leaf True) (Leaf False)
Node (Leaf (0b0,True)) (Leaf (0b1,False))

traverse :: Applicative f => (a -> f b) -> Tree n a -> f (Tree n b)

itraverse :: Applicative f => (Wrd n -> a -> f b) -> Tree n a -> f (Tree n b)

traverse1 :: Apply f => (a -> f b) -> Tree n a -> f (Tree n b)

itraverse1 :: Apply f => (Wrd n -> a -> f b) -> Tree n a -> f (Tree n b)

itraverse_ :: forall n f a b. Applicative f => (Wrd n -> a -> f b) -> Tree n a -> f ()

Since: 0.1.1

Zipping

zipWith :: (a -> b -> c) -> Tree n a -> Tree n b -> Tree n c

Zip two Vecs with a function.

izipWith :: (Wrd n -> a -> b -> c) -> Tree n a -> Tree n b -> Tree n c

Zip two Trees. with a function that also takes the elements' indices.

repeat :: SNatI n => a -> Tree n a

Repeat a value.

>>> repeat 'x' :: Tree N.Nat2 Char
Node (Node (Leaf 'x') (Leaf 'x')) (Node (Leaf 'x') (Leaf 'x'))

Universe

universe :: SNatI n => Tree n (Wrd n)

Get all Vec n Bool indices in Tree n.

>>> universe :: Tree N.Nat2 (Wrd N.Nat2)
Node (Node (Leaf 0b00) (Leaf 0b01)) (Node (Leaf 0b10) (Leaf 0b11))

QuickCheck

liftArbitrary :: forall n a. SNatI n => Gen a -> Gen (Tree n a)

liftShrink :: forall n a. (a -> [a]) -> Tree n a -> [Tree n a]